services/service_bandit_system/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "service-bandit-system"
version = "0.1.0"
description = "LinUCB contextual bandit for personalized music theme recommendations"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "clickhouse-sqlalchemy>=0.3.2",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "prometheus-client>=0.21.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "sqlalchemy>=2.0.46",
    "starlette-prometheus>=0.10.0",
    "structlog>=25.5.0",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.15.0",
    "ruff>=0.8.0",
]

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
"*" = ["*.sql"]
